pyproject.toml
[project]
name = "fasthtml-fastapi-template"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "alembic>=1.15.1",
    "fastapi[standard]>=0.115.11",
    "httpx>=0.28.1",
    "pydantic>=2.10.6",
    "pydantic-settings>=2.8.1",
    "python-fasthtml>=0.12.4",
    "sqlalchemy>=2.0.39",
    "sqlmodel>=0.0.24",
    "svcs>=25.1.0",
    "typer>=0.15.2",
    "url-normalize>=1.4.3",
]

[dependency-groups]
dev = ["pytest>=8.3.5", "ruff>=0.11.0"]


#########################################
# Linting and Formatting Configurations #
#########################################

[tool.ruff]
# Exclude a variety of commonly ignored directories.
extend-include = ["*.ipynb"]
exclude = [
    ".bzr",
    ".direnv",
    ".eggs",
    ".git",
    ".git-rewrite",
    ".hg",
    ".ipynb_checkpoints",
    ".mypy_cache",
    ".nox",
    ".pants.d",
    ".pyenv",
    ".pytest_cache",
    ".pytype",
    ".ruff_cache",
    ".svn",
    ".tox",
    ".venv",
    ".vscode",
    "__pypackages__",
    "_build",
    "buck-out",
    "build",
    "dist",
    "node_modules",
    "site-packages",
    "venv",
]

line-length = 88
indent-width = 4

# Assume Python 3.12
target-version = "py312"

[tool.ruff.lint]
# Ruff rules 
# https://docs.astral.sh/ruff/rules/
select = [
    # Defaults
    "E", # pycodestyle
    "F", # Pyflakes
    # Custom rules
    "W",    # pycodestyle warnings
    "C",    # McCabe complexity
    "B",    # flake8-bugbear
    "I",    # isort
    "UP",   # pyupgrade
    "YTT",  # flake8-2020
    "RUF",  # Ruff-specific rules
    "NPY",
    "FURB",
    "PERF",
    "ERA",
]
ignore = ["E501", "B904"]

# Allow fix for all enabled rules (when `--fix`) is provided.
fixable = ["ALL"]
unfixable = []

# Allow unused variables when underscore-prefixed.
dummy-variable-rgx = "^(_+|(_+[a-zA-Z0-9_]*[a-zA-Z0-9]+?))$"

[tool.ruff.format]
# Like Black, use double quotes for strings.
quote-style = "double"

# Like Black, indent with spaces, rather than tabs.
indent-style = "space"

# Like Black, respect magic trailing commas.
skip-magic-trailing-comma = false

# Like Black, automatically detect the appropriate line ending.
line-ending = "auto"

# Enable auto-formatting of code examples in docstrings. Markdown,
# reStructuredText code/literal blocks and doctests are all supported.
#
# This is currently disabled by default, but it is planned for this
# to be opt-out in the future.
docstring-code-format = false

# Set the line length limit used when formatting code snippets in
# docstrings.
#
# This only has an effect when the `docstring-code-format` setting is
# enabled.
docstring-code-line-length = "dynamic"

#########################
# Typing Configurations #
#########################

[tool.pyright]
exclude = ["**/__pycache__", ".venv", "**/alembic"]

typeCheckingMode = "strict"

# Setting explicitly defined from Pyright Configuration
# https://github.com/microsoft/pyright/blob/main/docs/configuration.md

# reportMissingTypeStubs = "none"
reportUnknownMemberType = "none"
reportUnknownArgumentType = "none"
# reportUnknownVariableType = "none"
reportCallInDefaultInitializer = "warning"
reportUnnecessaryTypeIgnoreComment = "warning"
reportShadowedImports = "warning"
reportPropertyTypeMismatch = "warning"
reportPrivateImportUsage = "none"
# reportConstantRedefinition = "none"
